Null-coalesce emoji lookup to false instead of muting
`@` prevents warnings from being printed, but it does not prevent them from being reported to debugger by Xdebug. Since most characters tend not to be emoji, the flood of notices sent to the debugger slows parsing to a crawl. Null-coalesce the element access to false instead.

(To think I almost discarded a perfectly good XLIFF library because I thought it was too :snail: :wink:)